backend/requirements.txt
APScheduler==3.11.2
bcrypt==3.2.2
email-validator==2.3.0
fastapi==0.110.1
motor==3.4.0
pymongo==4.6.1
pydantic==2.12.5
PyJWT==2.11.0
python-dotenv==1.2.1
reportlab==4.4.10
uvicorn==0.25.0